requirements.txt
streamlit
groq
requests
beautifulsoup4
lxml
python-dotenv
yfinance